Make HUBAPI_DOMAIN_OVERRIDE environment independent
Reworks getHubSpotApiOrigin to append the environment suffix to a passed environment variable HUBAPI_DOMAIN_OVERRIDE.
This makes the override independent to environments.
You can no longer mistakenly pass a domain override which conflicts with the environment of the account.
It also lets you "set and forget" the HUBAPI_DOMAIN_OVERRIDE variable in your shell settings, say in .zshrc, and it will work as expected in all cases.

Adds tests for urls.js to test the normal cases, as well as tests for the override behavior.